use .ocap.idle instead of polling from client
this currently requires the idle-callback-hack branch of Rserve, and is
not expected to be the final design.

with the hack, Rserve calls us whenever it is done processing its
own (ocap) messages through the file descriptors it is monitoring. @s-u
would prefer that we figure out what file descriptors Shiny is using,
and have Rserve monitor those as well. this will be easy if Shiny is
only using one port.